backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
schedule==1.2.0
python-multipart==0.0.6
aiohttp==3.9.1
selenium==4.15.2
webdriver-manager==4.0.1
psycopg2-binary==2.9.9  # PostgreSQL driver
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4  # Hash de senhas
python-multipart==0.0.6  # Para forms

